fixing indices
The GIN index for full-text search wasn't in the list of indices. I think the script needs to run, creating the materialised column, then the index needs to be created after the change commits, because splitting them across scripts works.

I made a small change to the indices because there was a typo and I'm a terrible pedant